Reworked the pages for relevant standards
The pages for relevant standards and documents for
HTTP-based APIs are now generated by a set of
shell scripts. These scripts generate the
content of the page by processing a YAML document,
containing all standards, with the help of jq.

In the result, to change the content of the
section with standards for HTTP-based API,
it is enough to add it to the YAML document.